% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowchart-render.R
\name{render_flowchart}
\alias{render_flowchart}
\title{Render a flow diagram for viewing}
\usage{
render_flowchart(x, background = "white")
}
\arguments{
\item{x}{A \code{reportilo_flowchart}.}

\item{background}{Diagram background color (default \code{"white"}; use
\code{"transparent"} for no background).}
}
\value{
A \code{DiagrammeR} \code{grViz} htmlwidget.
}
\description{
Render a \code{reportilo_flowchart} as an interactive Graphviz widget (for use in
the console, RStudio Viewer, R Markdown or Shiny). To save to a file, use
\code{reportilo_export()}.
}
\examples{
\dontshow{if (requireNamespace("DiagrammeR", quietly = TRUE)) withAutoprint(\{ # examplesIf}
fc <- new_flowchart("prisma_2020")
render_flowchart(fc)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{reportilo_export()}, \code{\link[=flowchart_dot]{flowchart_dot()}}
}
